+++ CHANGELOG.md
@@ -6,8 +6,9 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ## 0.0.4 - (Unreleased)
 ### Added
 - Checks for the following Quota's:
-  - Public IP Addresses
+  - Network Interfaces
   - Network Security Groups
+  - Public IP Addresses
   - Static Public IP Addresses
   - Virtual Networks
 ### Changed